Uses version 4.x of Google Terraform provider
This commit updates the Google Terraform Provider constraints to permit
any 4.x version to be used.

The Concourse Terraform Resource has been failing with

╷
│ Error: Provider produced inconsistent result after apply
│
│ When applying changes to google_service_account.opsman_service_account,
│ provider "provider[\"registry.terraform.io/hashicorp/google\"]" produced an
│ unexpected new value: Root resource was present, but now absent.
│
│ This is a bug in the provider, which should be reported in the provider's
│ own issue tracker.

The fault for this certainly doesn't lie in a Terraform Provider code
change. We've been using version 3.x of the Google Terraform provider.
The latest release of that is 3.90.1, which was released back in
November 2021.

Our hunch is that changes in GCP are causing the provider to fail, so
we're upgrading it, and making the required changes to the various
terraform files.

Signed-off-by: Joseph Palermo <jpalermo@vmware.com>